requirements.txt
PyQt5>=5.15.0
SQLAlchemy>=2.0.23
playwright>=1.40.0
requests>=2.31.0
Pillow>=10.2.0
openai>=1.3.7
python-dotenv>=1.0.0
colorama>=0.4.4
pydantic>=2.0.0
typing-extensions>=4.0.0
aiohttp>=3.8.0
